changing of design and home page

+++ frontend/src/app/issue-list/issue-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Issue } from '../models/Issue.model';
+import { DialogService } from '../services/dialog.service';
 import { IssuesService } from '../services/issues.service';
 
 @Component({
@@ -13,7 +14,11 @@ export class IssueListComponent implements OnInit, OnDestroy {
   issues: Issue[];
   issuesSubscription: Subscription;
 
-  constructor(private issuesService: IssuesService, private router: Router) {}
+  constructor(
+    private issuesService: IssuesService,
+    private router: Router,
+    private dialogService: DialogService
+  ) {}
 
   ngOnInit(): void {
     this.issuesSubscription = this.issuesService.issuesSubject.subscribe(
@@ -28,7 +33,15 @@ export class IssueListComponent implements OnInit, OnDestroy {
     this.router.navigate(['/issues', 'new']);
   }
   onDeleteIssue(issue: Issue) {
-    this.issuesService.removeIssue(issue);
+   
+    this.dialogService
+      .openConfirmDialog('Are you sure to delete this issue')
+      .afterClosed()
+      .subscribe(res => {
+       if(res) {
+        this.issuesService.removeIssue(issue);
+       }
+      });
   }
   onViewIssue(id: number) {
     this.router.navigate(['/issues', 'view', id]);
